Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/processing-model-2/runtime-error-in-window-onerror-expected.txt

CONSOLE MESSAGE: line 20: ReferenceError: Can't find variable: undefined_variable
CONSOLE MESSAGE: line 16: ReferenceError: Can't find variable: undefined_variable_in_onerror

PASS runtime error in window.onerror 
